Remove `Recents` from the Finder favourites
Firstly, `Recents` is the default, so you need to set it to be something else. I've set it to the home directory here; the name is found by googling defaults write finder new window.

Secondly, remove it using the script. However, it's not a directory but a file, and this may not work on all computers. On Steve's mac, it's `Recents.cannedsearch`, so different name and casing. If it doesn't work, print out what it's called in the script:

```
print sidebar.getNameFromIndex(2);
```

where the index is 0-based.