Configure fstab for my backup device
I have a My Book external hard drive that I use for backups, but by default when it's plugged in to a Mac it is Read-only. To make it writeable, I need to create this file in `/etc/fstab`, which requires downloading it and then `sudo mv`ing it. I can't seem to do the latter as part of `main.conf` though possibly I could if I read the code more closely.