Add a lock screen menu item
I like to be able to lock the screen with a mouse click. This used to be possible using Keychain Access, but that has been removed in favour of a click right down at the bottom of the Apple menu. @SteveMarshall wrote an Apple Script that does what I need, and this commit:

- Downloads that script and puts it in the ~/Library/Scripts folder (and also creates said folder if it doesn't exist)
- Adds the menu item to the menu bar
- Configures the menu item so it only shows our user script and not all the other library scripts