Remove Chrome and Firefox for now
The computer I just set up already had Chrome and Firefox installed and owned by another user so this caused `suited` to fail. Removing them for now so I can run `suited` again on this computer. Ideally `suited` wouldn't crash and burn with this, not sure if there's a way to fix that.