Add a note about checking for Firefox and Chrome
When I set up this computer, `suited` failed because of Firefox and Chrome already being installed and owned by someone else. For now I have removed them, but this commit adds a note and a link to the commit where I did so, so that next time I know to check and add them back in if they are not present.

Not ideal, but will do for the time being.