Python3 requires you install the bridge to obj-c
MacOS ships with a Python Objective-C bridge, to allow python to access and use the native MacOS Cocoa and Foundation frameworks, and this script relies on that bridge because it is manipulating things like the Finder.

However, MacOS ships Python 2x, and Python3 does not work with that bridge, so one needs to install one.

[This is a good article](https://scriptingosx.com/2020/02/wrangling-pythons/), which suggests *not* doing what I have done here, i.e. installing it for a single user on a single Mac. However, for my current purposes, I do not need to worry about multiple users or even about running multiple versions of Python myself right now, so going with it.